Fix test & adjust hierarchy.

+++ app/src/main/java/com/sloppylinux/mchl/util/MCHLService.java
@@ -1,6 +1,6 @@
 package com.sloppylinux.mchl.util;
 
-import com.sloppylinux.mchl.domain.Event;
+import com.sloppylinux.mchl.domain.retrofit.Event;
 import com.sloppylinux.mchl.domain.Player;
 import com.sloppylinux.mchl.domain.Team;
 
@@ -19,17 +19,17 @@ public interface MCHLService {
     Call<List<Team>> listTeams();
 
     @GET("teams/{team}")
-    Call<Team> getTeam(@Path(value = "team") int teamId);
+    Call<Team> getTeam(@Path(value = "team") long teamId);
 
     @GET("events")
     Call<List<Event>> listEvents();
 
     @GET("events/{event}")
-    Call<Event> getEvent(@Path(value = "event") int eventId);
+    Call<Event> getEvent(@Path(value = "event") long eventId);
 
     @GET("players")
     Call<List<Player>> listPlayers(@Query("per_page") Integer perPage, @Query("page") Integer page);
 
     @GET("players/{player}")
-    Call<Player> getPlayer(@Path(value = "player") int playerId);
+    Call<Player> getPlayer(@Path(value = "player") long playerId);
 }
